Message if a new revision is detected between
A bug, sort of, is found where information could get lost.

Scenario:
User A creates a new entity (Revision 1). The entity is saved, then user A want to fine-tune its position and starts to move the entity.

User B, from the same camp, opens uo the newly created entity (revison 1) to add a name and description. User B closes and saves the changes (revision 2).

Soon after User A has finished fine-tuneing and a save occurs it will save the first name and description from revision 1 and enter the new position change and completely ignore the newly given information from revision 2. All information given in revision 2 is now lost.

Attempted bug-fix:
Check if there's a revision in between the edit started when in the LayedDoneEdit-event.
If a in-between-revision is detected a change-log will be written in the description before saving the new revision. The user will also be notified about the between revision and asked to double-check the information.