Fix duplication bug after refresh
Bug:
When an entity was updated or added it was duplicated when going back to normal mode efter edit.
This duplicate appered after a refresh had been made.

Cause:
After an update the old entity's layer is removed, and the new enity is drawn. But this remove-function removed the position of the latest entity's in the repository.
When a refresh later is done it compares the list of old enity's with the fetched, and because the entity was removed it thinks it's a new entity and will add it.

Fix:
Set the flag to not remove the entity from the repository.